Cleansup and move tabs into components
