+++ src/components/spacer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-// Purpose of this component is to provide horizontal and vertical spacing whee possible instead of using margins
+// Purpose of this component is to provide horizontal and vertical spacing when possible instead of using margins
 const Spacer = ({ width, height }) => {
   return <div style={{ width: `${width}px`, height: `${height}px` }} />;
 };
