--- src/components/test.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Test = () => {
-  return <div>Yes</div>;
-};
-
-export default Test;
